tests/kms_chamelium: capture audio data in real-time
Before this patch, the audio test first sends an audio signal for 2s, and then
checks whether the captured signal matches.

This patch makes it so we send and check the signal in parallel. Thus we can
stop the test as soon as we receive the correct signal. This saves ~2s per
audio test.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>