tests/kms_chamelium: run audio test with multiple sampling rates
The audio test is now run multiple times with a variety of playback sampling
rates.

We now query the capture audio format from the Chamelium XML-RPC API instead of
hardcoding it.

One limitation is that we need to start sending an audio signal before being
able to query the capture audio format. However we need the capture sample rate
to decide which frequencies we generate. For now we use the playback rate and
check that it's the same as the capture rate.

Another limitation is that the DP receiver reports an unknown sampling rate
during the 44.1KHz test. In this case we assume the capture rate is the same as
the playback rate. We'll fail later anyway if this assumption is incorrect
since we check the signal we receive.

Chameleon bug: https://crbug.com/950913

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>