tests/kms_setmode: Handle eDP with fixed mode better.
When running the stealing tests, we set the same mode on the eDP and
the other connector. If the eDP is 4k, we may try to set a mode that
is not supported on the other connector, which might be 1080p.

Fix this by selecting the smallest default mode, based on clock.
Hopefully this is more likely to be supported on all connectors,
instead always using the first connector's mode.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=110391
[mlankhorst: Kill off 1 extra copy of mode (vsyrjala)]
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>