tests/kms_cursor_crc: Run tests once per pipe
Instead of running tests on all pipes and all outputs,
change the tests to only run once on each pipe, with
igt_get_single_output_for_pipe() to select an output.

This way we'll test all pipes and all outputs at least once,
but not all combinations. This reduces the amount of time
required for each test with dual/triple outputs.

In order to reduce the test time further, only run the
DPMS/suspend tests once for each pipe. They should not
be dependent on cursor size.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>