tests/kms_cursor_crc: Use igt_kms calls for everything.
Instead of doing direct drm calls, we can use the core to do it for
us. This can be used to increase logging verbosity, but it will also
allow us to test on overlay planes by changing all references from
DRM_PLANE_TYPE_CURSOR to DRM_PLANE_TYPE_OVERLAY.

This fails on gen9-10 because of the alpha rounding error bug, but it
can be useful for testing if the tests themselves are broken.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>