tests/kms_chamelium: Make sure we wait for each connectors' hotplug event
After scheduling an HPD toggle event, make sure that we wait for the
hotplug event for each connector that may be sent by the driver.

Depending on the scheduling there could be 1 event or as many events as
connectors we scheduled an HPD toggle event on, depending on the timing.
So if we don't yet see the expected connector state on a given connector
try to wait for an additional hotplug event and reprobe/recheck the
state.

v2:
- s/igt_assert(x >= y)/igt_assert_lte(y, x)/ to see the actual limits in
  the debugging output. (Lyude)

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=110534
Cc: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Cc: Lyude Paul <lyude@redhat.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Lyude Paul <lyude@redhat.com>
Reviewed-By: Paul Kocialkowski <paul.kocialkowski@bootlin.com>