lib/igt_alsa: support all alsa formats
This commit adds a new format argument to alsa_configure_output. This allows
the caller to decide the format used for the audio signal.

This removes the assumption that samples are S16_LE items in igt_alsa. The
alsa function snd_pcm_format_physical_width is used instead of sizeof(short).
This also removes the theorically incorrect assumption that sizeof(short) ==
sizeof(int16_t).

Right now S16_LE is hardcoded in the test. Tests with other formats will be
added in future commits.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>