lib: Show other clients when SETMASTER fails
Currently we only show the clients at DEBUG level which is not
automatically shown for a SKIP. However, failing to SETMASTER is more
often a mistake and we do want to see the other clients. The choice is
at what level to show them? If we want to avoid having extra WARNs in
CI, we should keep them to INFO, but on the other hand failing to acquire
SETMASTER is tantamount to failure, so WARN.

References: https://bugs.freedesktop.org/show_bug.cgi?id=110682
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Martin Peres <martin.peres@free.fr>
Acked-by: Martin Peres <martin.peres@free.fr>