meson: Start using 'feature' options
Meson 0.47 comes with a new type of option called 'feature' so instead of:
       type : 'combo',
       value : 'auto',
       choices : ['auto', 'true', 'false'],
We can:
       type : 'feature',

The main difference is that the feature takes auto, enabled and disabled
instead of auto, true and false.

get_option() on a feature returns opaque object that can be passed as
a 'required' argument of a dependency. Auto is equivalent to 'required
: false', enabled is equivalent to 'required : true' and disabled
introduces new behavior forcing the dependency to be considered not
found.

This allows us to streamline a lot of logic regarding optional IGT
features.

This patch bumps required meson version to 0.47.0

Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Petri Latvala <petri.latvala@intel.com>
Cc: Simon Ser <simon.ser@intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>