lib/igt_audio: fix amplitude for S24_LE and S32_LE
We need to use signed integers (hence "S" in the format name).

The reason this has gone unnoticed is that we were generating signals with half
of the amplitude they were supposed to have. As a result, using UINT32_MAX
(which is almost INT32_MAX * 2) resulted in the correct amplitude. In other
words, these two bugs were cancelling each other's effects.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Fixes: 42576d710255 ("lib/igt_audio: add support for S24_LE and S32_LE signal generation")