tests/kms_big_fb: Run the test for a single pipe only
I think we can safely assume that if no plane on the first
available pipe supports the format+mod+rotation combo we
don't have to check on all the other pipes.

The only slight exception is CHV where pipe B has a few more
features, but none of those are actually relevant for this
test.

Also leave put the PIPE_ANY commit as that still causes
pointless modesets.

$ time kms_big_fb
- real    2m8.101s
+ real    0m52.902s

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>