tests/kms_chamelium: add audio channel alignment test
The previous pulse test only checked the signal amplitude.

This commit adds a new check to the flatline test: channel alignment. This check
makes sure there is no time shift between each channel.

This is achieved by first sending a positive signal and then a falling edge.
Edges in each channel should be aligned.

The index of each channel's falling edge is stored in number of samples. I
chose not to implement a per-page test because the edge could be right between
two pages.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>