lib/igt_audio: detect noise and pops
First, normalize the bin power by dividing it by the number of input samples.
We need to multiply by 2 since we get half as many bins as input samples.

Second, check that low frequencies are under a given threshold. If there is a
pop or some noise, the low frequencies will be affected.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>