tests/kms_chamelium: reduce flatline test amplitude
Set flatline test amplitude to 0.1 instead of 0.9. There have been concerns
about edge-cases where bad TV hardware is connected via HDMI/DP and breaks when
a DC offset is sent.

This is a first mitigation. A second mitigation to only send audio to the
Chamelium device instead of all HDMI/DP ports is planned, but requires changes
in ALSA and/or DRM.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>