lib/i915: Use engine discovery in gem_test_engines
If request is to verify all engines work, use the new engine discovery API
so we can actually test all.

There is a sporadic and mysterious, possibly memory corruption issue
surronding this area so lets see if this changes the pattern.

It is easy to do in this function due it re-opening the driver and
throwing it away. So there are no concerns that the modified default
context can affect any other test.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
References: https://bugzilla.freedesktop.org/show_bug.cgi?id=110667
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>