tests/kms_plane: Throw away yet another bit
CHV pipe A/C sprites are causing a crc mismatch with BT.601
when we keep the six msbs. If we throw away one more bit we get
matches for BT.601. BT.709 matches even with 6 bits, as do the
pipe B planes with their programmable CSC. Also IVB and KBL
were both happy with 6 bits, so doesn't seem that that these
CHV mismatches are due to bugs in our code, just the hw is
a bit imprecise.

v2: s/bi/bit/ in the subject (Nicholas)

Cc: Uma Shankar <uma.shankar@intel.com>
Reviewed-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Reviewed-by: Uma Shankar <uma.shanka@intel.com>
Tested-by: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>