tests/i915/gem_spin_batch: Update with engine discovery
Legacy execbuf abi tests are prefixed with legacy. New test are added to
run on physical engines accessed through engine discovery.

So legacy tests run on the unconfigured (with engine map) context and
use a new helper gem_eb_flags_to_engine to look up the engine from the
intel_execution_engines2 static list. This is only to enable the
core test code to be shared.

Places where new contexts are created had to be updated to either
equally configure the contexts or not.

v2:
  helper called gem_engine_is_equal is added [tvrtko]
  rebased to adopt the changte in eb_flag to engine [tvrtko]
v3:
  returned the result of the expression itself [tvrtko]

Signed-off-by: Ramalingam C <ramalingam.c@intel.com>
Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>