lib/igt_kms: remove length parameter from kmstest_force_edid
Given an EDID, computing the size is trivial. Instead of having one size
constant per EDID and hope the callers use the right one (ie. *not* EDID_LENGTH
when there's an extension), we can make functions that take EDIDs compute the
size if they need it.

We have tests in lib/tests/igt_edid.c which assert the number of extensions
present in the EDID anyway.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>