lib: Ditch "DRIVER_* for kms tests" cargo cult
KMS is meant to be a (at least somewhat) generic userspace API. We do
not want nor need to add a special match function for every driver
ever written, that doesn't make sense, defeats the point of having
generic tests for a generic testsuite, and really doesn't scale.

Also add a comment so people don't try to add ever more of these.

Also, this means no autoloading for you, but really igt should
reinvent udev, and mostly this was needed for 2 reasons:
- CI configuration falling to pieces because tests unloaded a driver,
  and then died before cleaning up.
- vgem, which is fake, and needs to be requested explictily.

We might want to throw that all out again, except for vgem.

Cc: Liviu Dudau <liviu.dudau@arm.com>
Cc: Deepak Rawat <drawat@vmware.com>
Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Cc: Leo Li <sunpeng.li@amd.com>
Cc: Michał Winiarski <michal.winiarski@intel.com>
Cc: "Ser, Simon" <simon.ser@intel.com>
Cc: Oleg Vasilev <oleg.vasilev@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Acked-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Reviewed-by: Simon Ser <simon.ser@intel.com>
Acked-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>