tools/vbt_decode: Fix VBT parsing for the PSR section
A single 32-bit PSR2 training pattern field follows the sixteen element
array of PSR table entries as per VBT spec. But, we incorrectly define
this PSR2 field for each of PSR table entries. The result of this is
that the PSR1 training pattern duration for any panel_type != 0 will be
parsed incorrectly.

Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>